requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
dj_database_url
gunicorn
whitenoise
